modularisation and fixed types
